refactor(#792): extract build_audio_processor_for_model into its own file (sweep)
runtime_block_builders.rs 676 -> 519 (leaves the debt list). Moved the audio-mode
-> AudioProcessor resolution (build_audio_processor_for_model + expect_mono/stereo)
to runtime_processor_model.rs; the two callers import it directly. Engine tests
pass, zero warnings.
