+++ crates/engine/src/runtime_processor_model.rs
@@ -1,0 +1,176 @@
+//! Build an [`AudioProcessor`] for a block model from its audio-mode schema
+//! (issue #792 split — single responsibility).
+//!
+//! Given a model's declared audio mode and the input channel layout, this
+//! resolves the concrete mono / dual-mono / stereo processor shape (issue #588
+//! mono-collapse included) and validates the built `BlockProcessor` matches.
+//! Pure move out of `runtime_block_builders.rs`; the two callers
+//! (`runtime_block_core`, `runtime_block_builders::build_nam_audio_processor`)
+//! import `build_audio_processor_for_model` from here.
+
+use anyhow::{anyhow, Result};
+
+use block_core::{
+    AudioChannelLayout, BlockProcessor, ModelAudioMode, MonoProcessor, StereoProcessor,
+};
+use project::block::schema_for_block_model;
+use project::chain::Chain;
+
+use crate::runtime::layout_label;
+use crate::runtime_audio_frame::AudioProcessor;
+use crate::runtime_state::ProcessorBuildOutcome;
+
+pub(crate) fn build_audio_processor_for_model<F>(
+    chain: &Chain,
+    effect_type: &str,
+    model: &str,
+    input_layout: AudioChannelLayout,
+    content_mono: bool,
+    mut builder: F,
+) -> Result<ProcessorBuildOutcome>
+where
+    F: FnMut(AudioChannelLayout) -> Result<BlockProcessor>,
+{
+    let schema = schema_for_block_model(effect_type, model).map_err(|error| {
+        anyhow!(
+            "chain '{}' {} model '{}': {}",
+            chain.id.0,
+            effect_type,
+            model,
+            error
+        )
+    })?;
+
+    let output_layout = schema
+        .audio_mode
+        .output_layout(input_layout)
+        .ok_or_else(|| {
+            anyhow!(
+                "chain '{}' {} model '{}' with audio mode '{}' does not accept {} input",
+                chain.id.0,
+                effect_type,
+                model,
+                schema.audio_mode.as_str(),
+                layout_label(input_layout)
+            )
+        })?;
+
+    let processor = match (schema.audio_mode, input_layout) {
+        // MonoOnly: build mono processor — process_buffer handles stereo↔mono conversion
+        (ModelAudioMode::MonoOnly, _) => AudioProcessor::Mono(expect_mono_processor(
+            builder(AudioChannelLayout::Mono)?,
+            chain,
+            effect_type,
+            model,
+        )?),
+        (ModelAudioMode::DualMono, AudioChannelLayout::Mono) => {
+            AudioProcessor::Mono(expect_mono_processor(
+                builder(AudioChannelLayout::Mono)?,
+                chain,
+                effect_type,
+                model,
+            )?)
+        }
+        // Issue #588: a mono source broadcast to stereo carries identical
+        // channels (L == R). A single mono processor on that signal yields
+        // bit-identical output to two instances (`mono_mix([s, s]) == s`),
+        // at half the model footprint and CPU. Only collapse when the
+        // content reaching this block is still effectively mono.
+        (ModelAudioMode::DualMono, AudioChannelLayout::Stereo) if content_mono => {
+            AudioProcessor::Mono(expect_mono_processor(
+                builder(AudioChannelLayout::Mono)?,
+                chain,
+                effect_type,
+                model,
+            )?)
+        }
+        (ModelAudioMode::DualMono, AudioChannelLayout::Stereo) => AudioProcessor::DualMono {
+            left: expect_mono_processor(
+                builder(AudioChannelLayout::Mono)?,
+                chain,
+                effect_type,
+                model,
+            )?,
+            right: expect_mono_processor(
+                builder(AudioChannelLayout::Mono)?,
+                chain,
+                effect_type,
+                model,
+            )?,
+        },
+        (ModelAudioMode::TrueStereo, AudioChannelLayout::Stereo) => {
+            AudioProcessor::Stereo(expect_stereo_processor(
+                builder(AudioChannelLayout::Stereo)?,
+                chain,
+                effect_type,
+                model,
+            )?)
+        }
+        (ModelAudioMode::MonoToStereo, AudioChannelLayout::Mono) => {
+            AudioProcessor::StereoFromMono(expect_stereo_processor(
+                builder(AudioChannelLayout::Stereo)?,
+                chain,
+                effect_type,
+                model,
+            )?)
+        }
+        (ModelAudioMode::MonoToStereo, AudioChannelLayout::Stereo) => {
+            AudioProcessor::Stereo(expect_stereo_processor(
+                builder(AudioChannelLayout::Stereo)?,
+                chain,
+                effect_type,
+                model,
+            )?)
+        }
+        _ => {
+            return Err(anyhow!(
+                "chain '{}' {} model '{}' with audio mode '{}' cannot run on {} input",
+                chain.id.0,
+                effect_type,
+                model,
+                schema.audio_mode.as_str(),
+                layout_label(input_layout)
+            ));
+        }
+    };
+
+    Ok(ProcessorBuildOutcome {
+        processor,
+        output_layout,
+        stream_handle: None,
+    })
+}
+
+fn expect_mono_processor(
+    processor: BlockProcessor,
+    chain: &Chain,
+    effect_type: &str,
+    model: &str,
+) -> Result<Box<dyn MonoProcessor>> {
+    match processor {
+        BlockProcessor::Mono(processor) => Ok(processor),
+        BlockProcessor::Stereo(_) => Err(anyhow!(
+            "chain '{}' {} model '{}' returned stereo processing where mono was required",
+            chain.id.0,
+            effect_type,
+            model
+        )),
+    }
+}
+
+fn expect_stereo_processor(
+    processor: BlockProcessor,
+    chain: &Chain,
+    effect_type: &str,
+    model: &str,
+) -> Result<Box<dyn StereoProcessor>> {
+    match processor {
+        BlockProcessor::Stereo(processor) => Ok(processor),
+        BlockProcessor::Mono(_) => Err(anyhow!(
+            "chain '{}' {} model '{}' returned mono processing where stereo was required",
+            chain.id.0,
+            effect_type,
+            model
+        )),
+    }
+}
